Update to_s() method in Board class
I want the board to be printed in a prettier way, showing the border
lines in each cell rather than a 2D array. This makes for a better user
experience.

I updated the initial board to be full of ' ' instead of nil values for
better presentation of empty cells.

I broke up the to_s() method using the private row() helper method to
improve readability.