Create winner? method
Used the same method for both checking rows and columns to be DRY. Check
for diagonals is also its own method to reduce line length.

I also split the win_diags? method to be DRY.

I made the helper methods private since they are not needed outside of
a Board object.