Add logic for placing token in Game class
Thinking about it this could be in a Player class, but I decided to
commit since this logic took a while to write.

Includes input validation; so it must be in board range and on an
unoccupied cell.

Used a BOARD_LENGTH constant in the Board class, since this information
belongs to the Board and not anywhere else. Maybe this is pedantic, and
I have hard-coded the fact that the board is of length 3 elsewhere
anyway.

I also made an accessor for the board instance variable in Board, since
it is required to be both read and written outside of class instances.
This meant that I got rid of the add method in Board as I think it is
unnecessary extra code.