Add play turn functionality minus game over checks
I changed up the init method to have an array of player objects with
player_turn being the index position of the player whose turn it is.
This made it easier to change the player turn (just increment mod 2).

I also changed result instance variable to game_over, since this data is
more related to the Game class.

Note that I also changed @player_turn to @token to print the message to
the screen correctly in the Player choose_coord method.