Complete main file and make whole program work
Essentially the game now works!

The full? method was moved back to Board since it is a derived property
of board. Similarly the winner? method is moved to Player since it is a
derived property of the player (even though it uses Board info a lot).
The winner? method essentially replaces the win instance variable.

I moved the puts @board from the play_turn method in Game since this is
not strictly about playing a turn. It is now done in the main file.

The check_game_over method is complete, which changes the result message
if it finds an end game condition. Rather than having a separate
game_over instance variable, I use result.nil? to derive this.

I pass @board.board to the methods in Player instead of the board object
since the whole object is not needed.