Move loop from main file to Game class
This loop was logic for playing the game so it belongs in the Game
class.

Moved lines from play to play_turn method, to reduce play method size.
It also makes sense now that play_turn gets the current player to move,
checks for game over, switches the player turn, and prints the board;
this all counts as a turn I think.

Created current_player local variable and switch_player_turn method for
easier readability.

The check_game_over method now checks for game over before
switch_player_turn which allows us to check on the current player and
not look back on the previous player after switching. This makes the
logic easier to follow.

The switch_player_turn method uses a different way to switch,
(1 - index) with index either 0 or 1. This is easier to understand than
adding 1 modulo 2.