refactor!: make getting the current player public
This is to make testing easier, as current_player is now a variable and
not a private method. Now find_current_player can be tested as it is
public (part of the public script method #play_turn) separately from the
methods that used the #current_player before. These methods can now be
tested with ignorance of whose turn it is, and so follow the single
responsibility principle better.

Also ordered #find_current_player better: The order the methods are
declared matches the order the methods are called in the #play_turn
public script method.

BREAKING CHANGE: #make_move and #check_game_over now must take one
argument rather than none